# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: omu-rest
  description: 'Laravel製RESTful APIサーバー。著者: [connect0459](https://github.com/connect0459)'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api'
    description: ローカルホストのAPIエンドポイント
  -
    url: 'https://example.studyfortwo.org/omu-rest/public/api'
    description: lolipopのAPIエンドポイント
paths:
...
  /orders_payments:
    get:
      tags:
        - orders_payments
      summary: 注文支払いのリストを取得する
      operationId: 52140bdaf9965edfc692b030a9f1898f
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderPayment'
    post:
      tags:
        - orders_payments
      summary: 新しい注文支払いを作成する
      operationId: 231e4438ab66e1f3ba4205adf67e0df7
      requestBody:
        description: OrderPaymentのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPayment'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/OrderPayment'
  '/orders_payments/{id}':
    get:
      tags:
        - orders_payments
      summary: 特定のIDで注文支払いを取得する
      operationId: ed63e750658440b279ee8040dc02aba5
      parameters:
        -
          name: id
          in: path
          description: 注文支払いのID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPayment'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - orders_payments
      summary: 特定のIDで注文支払いを更新する
      operationId: e7043bcd7e8fc712b2e3a6516f1b8ca8
      parameters:
        -
          name: id
          in: path
          description: 注文支払いのID
          required: true
          schema:
            type: string
      requestBody:
        description: OrderPaymentのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPayment'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/OrderPayment'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - orders_payments
      summary: 特定のIDで注文支払いを削除する
      operationId: 41c8a438add65c3a1a6f1f7f0efeeea3
      parameters:
        -
          name: id
          in: path
          description: 注文支払いのID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
  /types_branches:
    get:
      tags:
        - types_branches
      summary: 支部のリストを取得する
      operationId: 97af74bfee220c89b0c0c5700fd29b99
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypeBranch'
    post:
      tags:
        - types_branches
      summary: 新しい支部を作成する
      operationId: ce050dbe8f9dcf8a1411581387496db6
      requestBody:
        description: TypeBranchのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeBranch'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/TypeBranch'
  '/types_branches/{id}':
    get:
      tags:
        - types_branches
      summary: 特定のIDで支部を取得する
      operationId: e33a8a3abd11b078e164c05b62246e43
      parameters:
        -
          name: id
          in: path
          description: 支部のID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeBranch'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - types_branches
      summary: 特定のIDで支部を更新する
      operationId: 98726087136b83ce1a215021e6eb76de
      parameters:
        -
          name: id
          in: path
          description: 支部のID
          required: true
          schema:
            type: string
      requestBody:
        description: TypeBranchのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeBranch'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/TypeBranch'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - types_branches
      summary: 特定のIDで支部を削除する
      operationId: 4f16c2a000161aa01c071b5beabaa1f2
      parameters:
        -
          name: id
          in: path
          description: 支部のID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
  /types_genres:
    get:
      tags:
        - types_genres
      summary: 書籍のジャンルのリストを取得する
      operationId: 6e1cdb613ad44750c97159198eac2c2e
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypeGenre'
    post:
      tags:
        - types_genres
      summary: 新しい書籍のジャンルを作成する
      operationId: 5876eb953745b47325ca0277455f8885
      requestBody:
        description: TypeGenreのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeGenre'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/TypeGenre'
  '/types_genres/{id}':
    get:
      tags:
        - types_genres
      summary: 特定のIDで書籍のジャンルを取得する
      operationId: 70509fad3bfe1e67a2caf35fe9ddcd88
      parameters:
        -
          name: id
          in: path
          description: 書籍のジャンルのID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeGenre'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - types_genres
      summary: 特定のIDで書籍のジャンルを更新する
      operationId: 32a2a32ec50d514a9cc3ca51fb2d6578
      parameters:
        -
          name: id
          in: path
          description: 書籍のジャンルのID
          required: true
          schema:
            type: string
      requestBody:
        description: TypeGenreのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeGenre'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/TypeGenre'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - types_genres
      summary: 特定のIDで書籍のジャンルを削除する
      operationId: 8ecb93fe6d9f820f901f57aaca0c8650
      parameters:
        -
          name: id
          in: path
          description: 書籍のジャンルのID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
...
components:
  schemas:
...
    OrderPayment:
      description: 'OrderPayment Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        order_customer_id:
          description: 外部キー制約（orders_customersテーブル）
          type: integer
          format: int64
          example: '1'
        numbering:
          description: 整理番号
          type: string
          format: string
          example: 231B001
        subtotal:
          description: 教科書代の小計
          type: integer
          format: int64
          example: '1000'
        postage:
          description: 送料
          type: integer
          format: int64
          example: '400'
        fee:
          description: 手数料
          type: integer
          format: int64
          example: '100'
        is_paid:
          description: 支払い済みか
          type: boolean
          format: boolean
          example: 'true'
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
    TypeBranch:
      description: 'TypeBranch Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        type:
          description: 支部名
          type: string
          format: string
          example: ○○大学
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
    TypeGenre:
      description: 'TypeGenre Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        type:
          description: ジャンル名
          type: string
          format: string
          example: ○○大学
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: api_token
      in: header